Updated UINavigationList to check if it has list item deletion enabled.
By default, when the user touches a Navigation List tablecell, it
triggers navigation. To enable the user to be able to delete cells in a
navigation list using UIDeleteTableCell, UINavigationList now checks for
the presences of the class "ui--show-delete-disclosures" on the
tableview. This class indicates that the user has enabled cell deletion.
In such a case, UINavigationList will not navigate to the next view,
allowing the user to select a cell by touching the circular indicators
on the left of the table cells.