Did a major refactor of navigation and the css function. Scaled up some icons.
Did a major refactor of $.UINavigationList. Before it was possible for
a user to bang on a navigation list, firing multiple events on the same
navigation item. This could result in the app trying to navigate beyond
the existing navigable elements, leaving the user with a blank screen.
I resolved this my putting a check at the end of each navigation
transition to check if there is a current view. If not, it resets the
navigation history and navigates back to the last valid view.

Refactored the Element.css() method. It now accepts an object of
property/value pairs like jQuery: $("tablecell").css({color:"red",
"font-weight: "bold", "background-color": "orange"});
"{}" are no longer required for passing a string argument. The old use
was:
$("tablecell").css("{color: red; font-weight: bold; background-color:
orange;}");
The new usage is:
$("tablecell").css("color: red; font-weight: bold; background-color:
orange;");
You can still pass it a single property/value pair to set a value:
$("tablecell").css("color", "red");
Passing a property only will return its value:
$("tablecell").css("color");
Some examples were updated to use the new argument style for css.

The switch control example was refactored to properly set the checked
and unchecked states of a switch.

I had to scale up some of the icons because their default size was
smaller than others.