Updated switch control, integrating new styles in to main styles
By default all switch controls resemble the ones in iOS 5. To get the
rectangular switches of the earlier versions of iOS, give the switch
controls a value of ui-kind="traditional".
The new switch controls don't' have labels but symbols for On and Off,
so no localization is necessary.
There's a third style for switch controls that you can get with the
attribute ui-kind="generic".
When you create a switch control dynamically, you need to pass the
value kind: traditional. This will create the default label markup for
that style. Otherwise, no labels are required for the newer style.