Added new tableview layout type: titled list
Finally resolved actionsheet alignment issue in horizontal orientation
in mobile Safari. There was no issue in standalone mode.
Added new type of tableview layout: title list. This is like the
tableview list used for songs. It has a secondary category for
alphabetical. This is only the structural groundwork. Still need to
implement the interaction of the alphabetical menu to scroll the
appropriate section into view.