ChocolateChip bumped to version 1.1.9
Added method "prepend". This inserts the content at the beginning of a
collection of nodes, just like prepend in jQuery.
Added method "append". This inserts the content at the end of a
collection of nodes, just like append in jQuery.
Added property "libraryName" with a value of "ChocolateChip" to
identify the variable "$" as belonging to ChocolateChip. This gets
checked at load time so if ChocolateChip is already loaded, it won't
load again, or if "$" is being used by some other library, then
ChocolateChip will alias to __$ and __$$.