Major update for ChocolateChip.js and reorganized file structure to match jQuery and Zepto versions.
Added new array methods:
eq: when used with an array of nodes return by $$, it returns the array index equal to the argument passed. Numbering starts at 0.
is: returns an array of all indexes matching the argument passed. Possible filters are tag, class or attribute.
has: returns an array of all indexes whose children match the argument. Possible filters are tag, class or attribute.
not: returns an array of all indexes not matching the argument passed. Possible filters are tag, class or attribute.

Added new element methods:
is: returns the element if it matches the argument passed.
has: returns  the element if it child nodes match the argument passed.
not: returns  the element if it does not match the argument passed.
prependTo: when executed on an array of nodes, it prepends them sequentially to the node passed as an argument.
appendTo: when executed on an array of nodes, it appends them sequentially to the node passed as an argument.
noop: a "no operation performed" or empty array for passing as an argument where a function is expected.
concat: takes the arguments and returns them as a single string. The argument can be an array of elements, or a comma separated series of strings passed as arguments.
childElements: returns the child nodes of an element. If no argument is passed, it returns all child nodes, otherwise it returns child nodes that match the argument passed. Possible filters are tag, class or attribute.
kids: alias for childElements.
siblings: returns the siblings of an element. If no argument is passed, it returns all siblings, otherwise it returns the siblings based on matches with the argument passed. Possible filters are tag, class or attribute.
closest: alias for ancestor, to match jQuery/Zepto function of the same name.