Update with new methods for ChocolateChip.js
Added method: hasNot. This is the opposite of has() and will return all elements whose children do not match the argument.
Refactored $.extend() to default to itself if only one argument  is passed. This imitates the behavior of the jQuery extend method.
Added short cut for Array.prototype.slice: $.slice().
Added methods for testing types: $.isArray(), $.isFunction() and $.isObject().
Added new method Element.html(). This is a shortcut for Element.empty().insert(). It replaces the element with the argument . If the argument is just text, it works the same as Element.text().
Fixed problem where Element. UIFlipSubview with top and bottom were actually the same. Now passing "top" flips the subview top forward-bottom up from behind. Passing "bottom" flips the subview bottom up forward and the top down behind.