Bumped ChocolateChip.js to version 1.3.6.
Removed use of Element.isSameNode and replaced with equality operatior '==='. This was throwing exceptions on Firefox because it has been deprecated in DOM level 4. Using the equality operator works the same.

Removed test for safari5 and safari5_1 since they weren't being used anywhere in the framework.

Updated iScroll initialization for UIActionSheet. The previous selector was incorrect.

Some minor updates to how a couple of the examples work.