Updated how $.processJSON method works in ChocolateChip.js
The prior behavior for $.processJSON was that it required that the json
file be assigned to a variable in the file itself. This was
inconvenient. Now by default it assigns the JSON object the the
variable 'data'. If that variable already exists or you are getting
several JSON objects, you can pass a second optional value for the
variable you want that JSON to be referenced by.