Major update to Ajax methods.
Removed dependency of xhr and xhrjson on an element in
ChocolateChip.js. They now belong to the $ object.
Added ajax() as alias to xhr() in ChocolateChip.js.
Updated xhr() to pass the request.responseText to the success callback
in ChocolateChip.js.
Updated xhr() to pass the response object to the error callback in
ChocolateChip.js.
Moved url to options object in xhr() method of ChocolateChip.js.
Fixed bug in concat where arrays of strings were not getting reduced
properly in ChocolateChip.js.
Fixed bug in text() where a value of '0' was not getting inserted due
to boolean cooercion in ChocolateChip.js.