Update for [].animate, [].hide, [].show, [].ready
Updated [].hide and [].show to handle empty arrays.
Fixed [].hide to properly set visibility.
Fixed [].animate to work with object of CSS property/values and with a
duration and easing as per docs.
Fixed [].ready to execute on any element, not just document.
Organized tests in tree to handle more types being added.