Can not use $.UIPaging within a $.UITabBar
My app includes a tab bar which has an article which itself contains a paging set of screens. Unfortunately when switching tabs the UITabBar removes the "current" and "previous" classes of all DOM elements, irrespective if they're a part of the tab bar or not.

This fix only removes the "current" and "previous" from all "article" and "nav" elements, therefore leaving the "section" elements inside the $.UIPaging element intact.

I'm sure there might be better way of doing this, but this has worked perfectly well for me so far.