Updated UISheet control to take IDs.
You can now pass the sheet’s ID to UISheetShow:
$.UISheetShow(‘#mySheet’). Using an ID to show allows you to have
multiple sheets in an app and only open the target one. Otherwise it
opens a sheet with the class “.sheet”. If there is only one sheet this
is fine. If you have more than one, you will need to use an assigned ID
to show it. The ID is provided at creation time:
$.UISheet({ id: “purchaseSheet”});